Whale data supporting documentation on labeling

whale.data_labels

whale.data_labels(assign_as = "param_table",
  file_location = "/data/column_names.rds", ...)

Arguments

assign_as
character passed as a name to be assigned to the created data object into the .GlobalEnv

file_location
the path to the file where the table is stored. If the file isn't found for whatever reason, this will become the location of the output file after the function has recreated the table.

Details

Parsing the naming convention rules from the Class A AIS Position Report Messages 1, 2, and 3 to help determing the column associations. Ultimately, we end up with the following possibilities.

Message ID

6 bits Identifier for this message 1, 2 or 3

Repeat indicator

2 bits Used by the repeater to indicate how many times a message has been repeated. See Section 4.6.1, Annex 2; 0-3; 0 = default; 3 = do not repeat any more.

User ID

30 bits MMSI number

Navigational status

4 bits 0 = under way using engine, 1 = at anchor, 2 = not under command, 3 = restricted maneuverability, 4 = constrained by her draught, 5 = moored, 6 = aground, 7 = engaged in fishing, 8 = under way sailing, 9 = reserved for future amendment of navigational status for ships carrying DG, HS, or MP, or IMO hazard or pollutant category C, high speed craft (HSC), 10 = reserved for future amendment of navigational status for ships carrying dangerous goods (DG), harmful substances (HS) or marine pollutants (MP), or IMO hazard or pollutant category A, wing in ground (WIG); 11 = power-driven vessel towing astern (regional use); 12 = power-driven vessel pushing ahead or towing alongside (regional use);13 = reserved for future use,14 = AIS-SART (active), MOB-AIS, EPIRB-AIS15 = undefined = default (also used by AIS-SART, MOB-AIS and EPIRB-AIS under test)

Rate of turnROTAIS

8 bits 0 to +126 = turning right at up to 708 deg per min or higher0 to -126 = turning left at up to 708 deg per min or higher Values between 0 and 708 deg per min coded by ROTAIS = 4.733 SQRT(ROTsensor) degrees per minwhere ROTsensor is the Rate of Turn as input by an external Rate of Turn Indicator (TI). ROTAIS is rounded to the nearest integer value.+127 = turning right at more than 5 deg per 30 s (No TI available)-127 = turning left at more than 5 deg per 30 s (No TI available)-128 (80 hex) indicates no turn information available (default).ROT data should not be derived from COG information.

SOG

10 bits Speed over ground in 1/10 knot steps (0-102.2 knots)1 023 = not available, 1 022 = 102.2 knots or higher

Position accuracy

1 bits The position accuracy (PA) flag should be determined in accordance with the table below:1 = high (<= 10 m)0 = low (> 10 m)0 = default

Longitude

28 bits Longitude in 1/10 000 min (+/-180 deg, East = positive (as per 2's complement), West = negative (as per 2's complement).181= (6791AC0h) = not available = default)

Latitude

27 bits Latitude in 1/10 000 min (+/-90 deg, North = positive (as per 2's complement), South = negative (as per 2's complement). 91deg (3412140h) = not available = default)

COG

12 bits Course over ground in 1/10 = (0-3599). 3600 (E10h) = not available = default. 3 601-4 095 should not be used

True heading

9 bits Degrees (0-359) (511 indicates not available = default)

Time stamp

6 bits UTC second when the report was generated by the electronic position system (EPFS) (0-59, or 60 if time stamp is not available, which should also be the default value, or 61 if positioning system is in manual input mode, or 62 if electronic position fixing system operates in estimated (dead reckoning) mode, or 63 if the positioning system is inoperative)

special maneuvre indicator

2 bits 0 = not available = default1 = not engaged in special maneuver2 = engaged in special maneuver(i.e.: regional passing arrangement on Inland Waterway)

Spare

3 bits Not used. Should be set to zero. Reserved for future use.

RAIM-flag

1 bits Receiver autonomous integrity monitoring (RAIM) flag of electronic position fixing device; 0 = RAIM not in use = default; 1 = RAIM in use. See Table

Communication state (see below)

19 bits See Rec. ITU-R M.1371-5 Table 49

Sync state (see below)

2 bits 0 UTC direct (sync from own integral GPS receiver)1 UTC indirect (own GPS unavailable - UTC sync from GPSreceiver on nearby ship or base station)2 Station is synchronized to a base station (base direct - GPSunavailable).3 Station is synchronized to another station based on thehighest number of received stations or to another mobilestation, which is directly synchronized to a base station (GPSunavailable)

Slot time-out

3 bits Specifies frames remaining until a new slot is selected0 means that this was the last transmission in this slot1-7 means that 1 to 7 frames respectively are left until slot change

Sub message

14 bits The sub message depends on the current value in slot time-out as described in Table 19

References

https://www.navcen.uscg.gov/?pageName=AISMessagesA

Examples

If the data is already loaded and associated with the assign_as name
whale.data_labels()
 "param_table already in working environment"

If the file exists, but table not in the environment
> whale.data_labels(file_location = "column_names.rds")
 "Table is availble in the session as param_table"

If it cant find the file path, it will parse, save and load into memory
> whale.data_labels(file_location = "column_names_missing.rds")
Table has been saved to column_names_missing.rds
Table is availble in the session as param_table

#> Error: <text>:1:4: unexpected symbol
#> 1: If the
#>        ^